# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Website UHS RDC API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/about.store": {
            "post": {
                "tags": [
                    "About"
                ],
                "summary": "Créer un nouvel About",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "objective",
                                    "phone",
                                    "email",
                                    "experience"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Qui sommes-nous ?"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Notre entreprise existe depuis..."
                                    },
                                    "objective": {
                                        "type": "string",
                                        "example": "Fournir un service de qualité"
                                    },
                                    "assignment": {
                                        "type": "string",
                                        "example": "Fournir un service de qualité"
                                    },
                                    "vision": {
                                        "type": "string",
                                        "example": "Fournir un service de qualité"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Kinshasa, RDC"
                                    },
                                    "instagram": {
                                        "type": "string",
                                        "example": "https://instagram.com/entreprise"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "example": "https://facebook.com/entreprise"
                                    },
                                    "twitter": {
                                        "type": "string",
                                        "example": "https://twitter.com/entreprise"
                                    },
                                    "linkedin": {
                                        "type": "string",
                                        "example": "https://linkedin.com/company/entreprise"
                                    },
                                    "youtube": {
                                        "type": "string",
                                        "example": "https://youtube.com/entreprise"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+243 999 888 777"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "contact@example.com"
                                    },
                                    "experience": {
                                        "type": "string",
                                        "example": "10 ans d'expérience"
                                    },
                                    "image1": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image2": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "bannier_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Création réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/about.update/{id}": {
            "post": {
                "tags": [
                    "About"
                ],
                "summary": "Mettre à jour un About",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'élément à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "objective": {
                                        "type": "string"
                                    },
                                    "assignment": {
                                        "type": "string"
                                    },
                                    "vision": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "instagram": {
                                        "type": "string"
                                    },
                                    "facebook": {
                                        "type": "string"
                                    },
                                    "twitter": {
                                        "type": "string"
                                    },
                                    "linkedin": {
                                        "type": "string"
                                    },
                                    "youtube": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "experience": {
                                        "type": "string"
                                    },
                                    "image1": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image2": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "bannier_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "404": {
                        "description": "Ressource introuvable"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/objective.store": {
            "post": {
                "tags": [
                    "Objectifs"
                ],
                "summary": "Create",
                "description": "Creation",
                "operationId": "createObjective",
                "requestBody": {
                    "description": "Enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation"
                                ],
                                "properties": {
                                    "designation": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "ysteme"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/message.delete/{id}": {
            "delete": {
                "tags": [
                    "Message"
                ],
                "summary": "Supprimer un message",
                "operationId": "destroyMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "supprimé avec succès"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}